+++ src/OpenTelemetry/Trace/CompositeActivityProcessor.cs
@@ -107,7 +107,7 @@ namespace OpenTelemetry.Trace
                 }
                 else
                 {
-                    var timeout = (long)timeoutMilliseconds - sw.ElapsedMilliseconds;
+                    var timeout = timeoutMilliseconds - sw.ElapsedMilliseconds;
 
                     if (timeout <= 0)
                     {
@@ -143,7 +143,7 @@ namespace OpenTelemetry.Trace
                 }
                 else
                 {
-                    var timeout = (long)timeoutMilliseconds - sw.ElapsedMilliseconds;
+                    var timeout = timeoutMilliseconds - sw.ElapsedMilliseconds;
 
                     // notify all the processors, even if we run overtime
                     result = cur.Value.Shutdown((int)Math.Max(timeout, 0)) && result;
